gentle delete messages

+++ utilites/del_messages.py
@@ -1,4 +1,4 @@
-from . import bot, time, types, DELETE_MESSAGES_TIMER
+from . import bot, time, types, DELETE_MESSAGES_TIMER, datetime, timedelta
 
 timer = time()
 del_messages_list = []
@@ -8,18 +8,36 @@ def del_messages_append(message: types.Message):
     """
     appends (message chat id, message id) to list
     """
-    del_messages_list.append((message.chat.id, message.message_id))
+    del_messages_list.append(
+        (
+            message.chat.id,
+            message.message_id,
+            datetime.today()
+            )
+            )
 
 
 async def del_messages():
     """
     Remove messages from chat by message_id, chat_id from list.
     """
-    for chat_id, msg_id in del_messages_list:
+    young_msg = []
+    now = datetime.today()
+
+    for chat_id, msg_id, msg_time in del_messages_list:
         try:
-            await bot.delete_message(chat_id, msg_id)
+
+            if now - timedelta(hours=1) > msg_time:
+                await bot.delete_message(chat_id, msg_id)
+            else:
+                young_msg.append((chat_id, msg_id, msg_time))
+
         except Exception:
             print("no message")
+
+    if young_msg:
+        del_messages_list.clear()
+        del_messages_list.extend(young_msg)
 
 
 async def del_msg_by_timer():
